Code for production recommender
This file provides the code for the production recommender system. Based on the data in the results spreadsheet, this code uses the approach presented in experiment two (10 nearest neighbours). A userId and movieId are provided as inputs and one movie rating is predicted for that combination. The score implies whether the user specified should view the movie or not. Data is not written back to the database.