Code for setup
The Python code in this file is used with the SQL code in the mlr.sql file to prepare the database for experiments used to develop the recommender system. Some code here is executed before SQL code, while some is executed after. The processes here setup a copy of the rating data that is used for testing while not destroying the original values.